fix: seperate mermaid and vegaLite's types from types.ts
