+++ src/components/visualization/vegaLiteViz/index.ts
@@ -2,5 +2,5 @@ import './vegaLiteViz.css'
 
 import VegaLiteViz from './vegaLiteViz'
 
-export * from '../../types'
+export * from './vegaLiteViz.types'
 export default VegaLiteViz
